+++ gatsby-config.js
@@ -111,6 +111,19 @@ module.exports = {
       options: {
         trackingId: "UA-1322093-4",
       },
-    },
+    },{
+      resolve: `gatsby-plugin-manifest`,
+      options: {
+        name: "Nikita",
+        short_name: "Nikita",
+        start_url: "/",
+        background_color: "#ffffff",
+        theme_color: "#105859",
+        display: "minimal-ui",
+        icon: "src/images/logo.png", // This path is relative to the root of the site.
+      },
+    },{
+      resolve: `gatsby-plugin-offline`
+    }
   ]
 }